ｋ

+++ Buttle.php
@@ -250,7 +250,7 @@ $Skill3Image4 = getDB1('select Skill3Image from Characters where id=?',[$id4]);
             enemyHP.innerHTML -= allPower * (100 / DlimitDebuff);
 			var HPBar = document.getElementById('HPBar');
 			console.log(HPBar);
-			HPBar.width = (parseInt(HPBar.width) * (enemyHP / enemyMaxHP)) +"px";
+			HPBar.width = (parseFloat(HPBar.width) * (enemyHP / enemyMaxHP)) +"px";
             console.log(allPower * (100 / DlimitDebuff));
             if (enemyHP.innerHTML <= 0) {
                 location.href = "./Result_Win.html";
